Add LLMs module using grafana-llm-app
This commit adds functionality that can be used to make requests
to LLMs via the grafana-llm-app plugin. The initial commit just
adds support for OpenAI and doesn't make any attempt to abstract
over more than one LLM provider. It includes a function which can
be used to stream chat completion results back to the caller.

Very much WIP, especially the export structure which I can't seem
to figure out. I'd like it if users could do something like...

    import { openai } from '@grafana/experimental/llms';

but I'm not sure if that requires a change to rollup?