Support function calling; add more helper functions
This commit enables support for handling function calling by having
'chatCompletions' and 'streamChatCompletions' default to returning the
entire chat completions response in the Observable, so that users can
extract the 'function_call' object if they're using function calls.

It also improves the docs on a ton of interfaces and functions since
they're now exposed to users.