Improve health check to return more granular details
The Dashboard AI team have requested that the plugin health check
returns more details about whether the OpenAI functionality is
configured, and whether that configuration is valid. This PR adds it,
and similar functionality for the vector service.

It's backwards incompatible with past health checks unfortunately, but
the llmclient package does include a change which should be compatible
with both the current and previous version of the plugin.

We're definitely due a refactor of the way we handle OpenAI config and
requests, but I think that will be part of #80.