Cache result of health check on app struct
The app is recreated by the plugin management system if settings
change, so the cached value will be discarded when the user updates
the settings to fix the health check error.

Still unsure if this is an ideal way of doing things, but it
speeds up health checks significantly while saving on OpenAI costs.