fix: make sure the Grafana OpenAI proxy works when hosted at a subpath
Prior to this commit we ignored the path of the configured OpenAI API
when modifying the proxied URL of a request, so that even if the OpenAI
API URL was set to something like "https://foo.com/llm" we would still
proxy requests to "https://foo.com".

This commit preserves the path of the configured URL so that subpaths
are honoured in the outgoing proxied request from the plugin.

I also fixed a missing early return after failing to modify the URL.

Fixes https://github.com/grafana/incident/issues/5107.