Abstract models to allow more providers
Rather than having the client pass in OpenAI specific models create an
abstraction for a default and a high accuracy model. For backwards
compatability gpt-3.5-turbo is mapped to default and gpt-4 is mapped to
high accuracy.

As part of this it is also required to read the body in order to
transform from an abstracted model to the specific model (or deployment
in the case of Azure) being used. A simple interface is used that
currently only supports Models and ChatCompletions, but can be expanded
to support StreamChatCompletions in the future (and refactor the
streaming code). This also means we now limit the endpoints that our API
allows to just be Models and ChatCompletions as well. These are all that
we have really seen used so it should be ok.