Remove preload: true from plugin.json
This causes Grafana to preload the plugin code whenever it starts
up rather than when the plugin is accessed, which slows startup
times. We don't need it since we don't offer any plugin extensions
or anything that requires the frontend to be loaded.

Note that this doesn't affect the availability of the backend, which
is always available.