Use non-crypto UUID generation for stream request
crypto.randomUUID() is only available if the Grafana instance is using
HTTPS, which may not be the case for many on-prem users who do not want
to set up certs when running in a local network. Using the uuid module
instead will allow the stream completions to still work in these
environments, and there is nothing that requires crypto for creating a
path.