feat: run MCP server in plugin backend using Grafana Live
This plugin starts up an MCP server using a custom transport based on
Grafana Live. This is similar to the SSE transport, but rather than
using a long-lived SSE connection to return events, the MCP server
returns them on a Grafana Live subscription. Similarly, rather than
accepting JSON-RPC messages on the `/messages` endpoint, we accept
messages over the `PublishStream` Grafana Live handler. This gets us
similar functionality to the SSE transport, including all those which
require statefulness.

The tools are imported directly from the [Grafana MCP repository][mcp-grafana]
so will be up-to-date with the latest version of the MCP server.

The original PR included some additional changes to the frontend to make
it easier for users to use the MCP functionality. These changes have
been split out into a separate PR to make this one easier to review.

[mcp-grafana]: https://github.com/grafana/mcp-grafana