fix: use public API to publish, where possible
https://github.com/grafana/grafana-llm-app/pull/601 added a hacky
solution to publish messages over the Grafana Live Websocket
connection rather than the HTTP API, making it work better in
HA setups, but it accesses internal APIs so is not a long term
solution.

Fortunately, https://github.com/grafana/grafana/pull/102325 has
now been merged and will be available in Grafana >=11.6.0, so we
can use the public API for all future versions of Grafana and only
require the internal API for older versions.

This commit does that, using the public API when the `publish`
function has the new signature (3 arguments) and only falling
back to accessing the internal APIs if that isn't available.