feat: refresh Grafana access token before expiry in Cloud
Prior to this commit, the Grafana Live session handler would not refresh
the access token before it expired. This causes issues when using long
running streams in Grafana, as the token would expire halfway through
the stream and the client would experience strange errors.

This commit adds a timer to the Grafana Live session handler to refresh
the access token one minute before it expires. On refresh, the timer
will reset to fire one minute before the token is due to expire again.