refactor: don't manually refresh tokens, use exchange client every time
Rather than storing an access and ID token in the Grafana Live session,
re-obtain them every time by either going through the token exchange
client (for the access token) or using the one provided in the request
by Grafana (for the ID token).

To make it clear that we don't have one we can use, remove the access token
from the GrafanaLiveContextFunc signature, so that it's clearer that it
isn't automatically provided and should be refreshed each time.

Note that the token exchange client provided by the authn package already
handles caching and automatic refreshing of tokens, so we don't need to
deal with refreshing them ourselves.

This should have no functional impact, but will make it easier to add
token refreshing to other places (such as #691).