refactor: add MCP struct and simplify access token usage
This commit adds a new MCP struct that encapsulates the MCP server
and Grafana Live server, as well as holding all MCP-related settings.
Finally, it adds a new access token client that encapsulates the
access token exchange process, including skipping it if we are
not in Grafana Cloud.

Rather than manually refreshing access tokens, this relies on the
fact that the authlib library's TokenExchangeClient will automatically
cache and refresh the access token if it is expired.

Note: the restructuring of the MCP side is partly motivated by wanting
to also use the access token client in the HTTP server in #691.