fix: handle case when nested provider differs
Some customers are reporting that their Azure OpenAI settings don't
seem to work, with the health check returning 404. It looks like this
is happening because while the nested 'provider' field is set to 'azure',
the root one is set to 'openai', so our backend doesn't create the correct
type of client.

This should fix https://github.com/grafana/support-escalations/issues/16670#issuecomment-2949047284.