Use Streamable HTTP endpoints for @grafana/llm components
And deprecate the Grafana Live transport, which has been difficult to debug
due to implicit statefulness.

This commit also lets users customise the app plugin used for the MCP
server, and hence allows them to host the MCP server with a custom set
of tools on their own app plugin.