chore: use publint and attw to verify package.json; improve
publint is a tool that checks the package.json file for common errors
with package publishing, such as whether distributed files are included
in the `exports` field, etc.

attw checks that types are exported correctly.

This commit adds both of these to the project under the `lint` script,
and adds a new `.attw.json` file to configure attw to not fail for
Node 10 bundling.