fix: correctly handle model conversion for Anthropic if no mapping is provided
Prior to this commit we had a constant mapping of 'base' / 'large' to
OpenAI model names. This was incorrect if the user chooses Anthropic
as the provider but doesn't override the model settings.

Instead, we now parameterize the mapping based on the provider.

We also handle partial mapping overrides much better; previously we
would fall back to the configured default model if the user didn't provide
a mapping for a given model, but now we just use the default. For
example, if the user configured a mapping for 'base' but not 'large'
and made a request with 'large', we would have fallen back to the
'base' model (since it was marked as 'default'). Now we use the
fallback default model instead.

Finally, we update the UI to make it more obvious when mappings have
been set (using placeholders), and switch based on the provider, too.

Fixes #741.
Fixes #645.