fix: use higher max tokens for health check
It turns out that setting max_completion_tokens causes the request to
_fail_ if the number of tokens to be generated is greater than the
limit, as opposed to max_tokens which just truncated the response.
This causes the health check to fail if it's too low.

This commit gets rid of the max_completion_tokens set in the health
check so that it can't fail for that reason.