fix: move llmclient back to root
As described in the [Go Module docs][docs], when a package lives in
a non-root directory, the module versions are discovered using tags
like <prefix>/<version> where 'prefix' is the location of the
package in the repo. It's much nicer sticking with tags like
'llmclient/v0.2.0' so let's just move the client back here.

[docs]: https://go.dev/wiki/Modules#publishing-a-release